src/dev_dashboard/requirements.txt
pyzmq==22.0.3
flatbuffers==1.12
pysimplegui==4.34.0
